+++ git-diff/tests/diff.rs
@@ -2,4 +2,4 @@ pub type Result<T = ()> = std::result::Result<T, Box<dyn std::error::Error>>;
 
 pub use git_testtools::hex_to_id;
 
-mod visit;
+mod tree;
